+++ pylatex/math.py
@@ -10,6 +10,50 @@ from .base_classes import Command, Container, Environment
 from .package import Package
 
 
+class Alignat(Environment):
+    """Class that represents a aligned equation environment."""
+
+    #: Alignat environment cause compile errors when they do not contain items.
+    #: This is why it is omitted fully if they are empty.
+    omit_if_empty = True
+    packages = [Package('amsmath')]
+
+    def __init__(self, aligns=2, numbering=True, escape=False):
+        """
+        Parameters
+        ----------
+        aligns : int
+            number of alignments
+        numbering : bool
+            Whether to number equations
+        escape : bool
+            if True, will escape strings
+        """
+        self.aligns = aligns
+        self.numbering = numbering
+        self.escape = escape
+        if not numbering:
+            self._star_latex_name = True
+        super().__init__(
+            start_arguments=[str(int(aligns))]
+        )
+
+    def add_math(self, data, **kwargs):
+        """Add math to the list.
+
+        Args
+        ----
+        data: str, `~.LatexObject`, list[str, `~.LatexObject`]
+            The equation itself.
+        kwargs : dict
+            any keyword arguments to Math container
+        """
+        escape = kwargs.pop('escape', None) or self.escape
+        if not isinstance(data, (list, tuple)):
+            data = [data]
+        self.append(Math(data=list(data), raw=True, escape=escape, **kwargs))
+
+
 class Math(Container):
     """A class representing a math environment."""
 
@@ -17,17 +61,24 @@ class Math(Container):
 
     content_separator = ' '
 
-    def __init__(self, *, inline=False, data=None):
-        """
+    def __init__(self, *, inline=False, data=None, raw=False,
+                 escape=False):
+        r"""
         Args
         ----
         data: list
             Content of the math container.
         inline: bool
             If the math should be displayed inline or not.
+        raw : bool
+            if raw, then will not be wrapped into environment
+        escape : bool
+            if True, will escape strings
         """
 
         self.inline = inline
+        self.raw = raw
+        self.escape = escape
         super().__init__(data=data)
 
     def dumps(self):
@@ -38,10 +89,12 @@ class Math(Container):
         str
 
         """
-
-        if self.inline:
+        if self.raw:
+            return self.dumps_content()
+        elif self.inline:
             return '$' + self.dumps_content() + '$'
-        return '\\[%\n' + self.dumps_content() + '%\n\\]'
+        else:
+            return '\\[%\n' + self.dumps_content() + '%\n\\]'
 
 
 class VectorName(Command):