+++ docs/source/changelog.rst
@@ -20,6 +20,9 @@ Fixed
 - Allow mappers of `~.dumps_list` to return a `~.LatexObject`.
 - Section numbering default behaviour fixed
 - Setter method for `~.LatexObject.escape` property added
+- Escape and raw flags to `.Math` container
+- `.LatexObject._star_latex_name` flag to append a star
+- `.Alignat` environment that can contain strings and `.Math` containers
 
 1.1.1_ - `docs <../v1.1.1/>`__ - 2016-12-10
 -------------------------------------------